requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
code-ast==0.1.0
code-tokenize==0.2.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.40
gunicorn==21.2.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
loguru==0.7.2
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
odfpy==1.4.1
openai==0.28.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
pinecone-client==2.2.4
plotly==5.18.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.14.1
torch==2.1.0+cpu
torchaudio==2.1.0+cpu
torchvision==0.16.0+cpu
tqdm==4.66.1
traitlets==5.13.0
transformers==4.34.1
tree-sitter==0.20.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
